Migrate ChipField
